Refactory to finished multilabel
