+++ classifip/models/mlc/nccbr.py
@@ -1,7 +1,8 @@
-from ..dataset.arff import ArffFile
-from ..representations.voting import Scores
+from classifip.dataset.arff import ArffFile
+from classifip.representations.voting import Scores
 import numpy as np
 from math import exp
+
 
 class NCCBR(object):
     """NCCBR implements the naive credal classification method using the IDM for
@@ -21,20 +22,19 @@ class NCCBR(object):
     :type feature_values: dictionnary associating each feature name to a list
     
     """
-    
-    
+
     def __init__(self):
         """Build an empty NCCBR structure
         """
-        self.feature_names=[]
-        self.label_names=[]
-        self.feature_values=dict()
-        self.feature_count=dict()
-        self.nblabels=0
-        self.trainingsize=0
-        self.labelcounts=[]
-        
-    def learn(self,learndataset,nblabels):
+        self.feature_names = []
+        self.label_names = []
+        self.feature_values = dict()
+        self.feature_count = dict()
+        self.nblabels = 0
+        self.trainingsize = 0
+        self.labelcounts = []
+
+    def learn(self, learndataset, nblabels):
         """learn the NCC for each label, mainly storing counts of feature/label pairs
         
         :param learndataset: learning instances
@@ -43,32 +43,29 @@ class NCCBR(object):
         :type nblabels: integer
         """
         self.__init__()
-        self.nblabels=nblabels
-        self.trainingsize=len(learndataset.data)
-        #Initializing the counts
-        self.feature_names=learndataset.attributes[:-self.nblabels]
-        self.label_names=learndataset.attributes[-self.nblabels:]
-        self.feature_values=learndataset.attribute_data.copy()
+        self.nblabels = nblabels
+        self.trainingsize = len(learndataset.data)
+        # Initializing the counts
+        self.feature_names = learndataset.attributes[:-self.nblabels]
+        self.label_names = learndataset.attributes[-self.nblabels:]
+        self.feature_values = learndataset.attribute_data.copy()
         for label_value in self.label_names:
-            label_set_one=learndataset.select_col_vals(label_value,['1'])
+            label_set_one = learndataset.select_col_vals(label_value, ['1'])
             self.labelcounts.append(len(label_set_one.data))
-            label_set_zero=learndataset.select_col_vals(label_value,['0'])
+            label_set_zero = learndataset.select_col_vals(label_value, ['0'])
             for feature in self.feature_names:
-                count_vector_one=[]
-                count_vector_zero=[]
-                feature_index=learndataset.attributes.index(feature)
-                for feature_value in learndataset.attribute_data[feature]:                   
-                    nb_items_one=[row[feature_index] for row in label_set_one.data].count(feature_value)
+                count_vector_one = []
+                count_vector_zero = []
+                feature_index = learndataset.attributes.index(feature)
+                for feature_value in learndataset.attribute_data[feature]:
+                    nb_items_one = [row[feature_index] for row in label_set_one.data].count(feature_value)
                     count_vector_one.append(nb_items_one)
-                    nb_items_zero=[row[feature_index] for row in label_set_zero.data].count(feature_value)
+                    nb_items_zero = [row[feature_index] for row in label_set_zero.data].count(feature_value)
                     count_vector_zero.append(nb_items_zero)
-                self.feature_count[label_value+'|in|'+feature]=count_vector_one
-                self.feature_count[label_value+'|out|'+feature]=count_vector_zero
-                
-            
-            
-        
-    def evaluate(self,testdataset,ncc_epsilon=0.001,ncc_s_param=2):
+                self.feature_count[label_value + '|in|' + feature] = count_vector_one
+                self.feature_count[label_value + '|out|' + feature] = count_vector_zero
+
+    def evaluate(self, testdataset, ncc_epsilon=0.001, ncc_s_param=2.0):
         """evaluate the instances and return a list of probability intervals.
         
         :param testdataset: list of input features of instances to evaluate
@@ -78,7 +75,7 @@ class NCCBR(object):
         :type ncc_espilon: float
         :param ncc_s_param: s parameter used in the IDM learning (settle
         imprecision level)
-        :type ncc_s_param: list
+        :type ncc_s_param: float
         :returns: for each value of ncc_s_param, a set of scores for each label
         :rtype: lists of :class:`~classifip.representations.voting.Scores`
  
@@ -98,47 +95,43 @@ class NCCBR(object):
             * solve the zero division problem
             
         """
-        #computing label proportions
-        label_prop=[n/float(self.trainingsize) for n in self.labelcounts]
-        answers=[]
+        # computing label proportions
+        label_prop = [n / float(self.trainingsize) for n in self.labelcounts]
+        answers = []
         for item in testdataset:
- 
-            #initializing scores
-            resulting_score=np.zeros((self.nblabels,2))
-            #computes product of lower/upper prob for each class
+
+            # initializing scores
+            resulting_score = np.zeros((self.nblabels, 2))
+            # computes product of lower/upper prob for each class
             for j in range(self.nblabels):
-                u_numerator=1-label_prop[j]
-                l_numerator=1-label_prop[j]
-                u_denom=label_prop[j]
-                l_denom=label_prop[j]
+                u_numerator = 1 - label_prop[j]
+                l_numerator = 1 - label_prop[j]
+                u_denom = label_prop[j]
+                l_denom = label_prop[j]
                 for f_index, feature in enumerate(self.feature_names):
-                    #computation of denominator (label=1)
-                    f_val_index=self.feature_values[feature].index(item[f_index])
-                    count_string=self.label_names[j]+'|in|'+feature
-                    num_items=float(sum(self.feature_count[count_string]))
-                    lower=(self.feature_count[count_string][f_val_index]/
-                            (num_items+ncc_s_param))
-                    l_denom=l_denom*((1-ncc_epsilon)*lower+
-                            ncc_epsilon/len(self.feature_count[count_string]))
-                    upper=((self.feature_count[count_string][f_val_index]+ncc_s_param)/(num_items+ncc_s_param))
-                    u_denom=u_denom*((1-ncc_epsilon)*upper+
-                            ncc_epsilon/len(self.feature_count[count_string]))
-                    
-                    #computation of numerator (label=0)
-                    count_string=self.label_names[j]+'|out|'+feature
-                    num_items=float(sum(self.feature_count[count_string]))
-                    lower=((self.feature_count[count_string]
-                            [f_val_index]+ncc_s_param)/(num_items+ncc_s_param))
-                    l_numerator=l_numerator*((1-ncc_epsilon)*lower+
-                            ncc_epsilon/len(self.feature_count[count_string]))
-                    upper=(self.feature_count[count_string]
-                            [f_val_index])/(num_items+ncc_s_param)
-                    u_numerator=u_numerator*((1-ncc_epsilon)*upper+
-                            ncc_epsilon/len(self.feature_count[count_string]))
-                resulting_score[j,1]=u_denom/(u_denom+u_numerator)
-                resulting_score[j,0]=l_denom/(l_denom+l_numerator)
-            result=Scores(resulting_score)
+                    # computation of denominator (label=1)
+                    f_val_index = self.feature_values[feature].index(item[f_index])
+                    count_string = self.label_names[j] + '|in|' + feature
+                    num_items = float(sum(self.feature_count[count_string]))
+                    lower = (self.feature_count[count_string][f_val_index] / (num_items + ncc_s_param))
+                    l_denom = l_denom * ((1 - ncc_epsilon) * lower +
+                                         ncc_epsilon / len(self.feature_count[count_string]))
+                    upper = ((self.feature_count[count_string][f_val_index] + ncc_s_param) / (num_items + ncc_s_param))
+                    u_denom = u_denom * ((1 - ncc_epsilon) * upper +
+                                         ncc_epsilon / len(self.feature_count[count_string]))
+
+                    # computation of numerator (label=0)
+                    count_string = self.label_names[j] + '|out|' + feature
+                    num_items = float(sum(self.feature_count[count_string]))
+                    lower = ((self.feature_count[count_string][f_val_index] + ncc_s_param) / (num_items + ncc_s_param))
+                    l_numerator = l_numerator * ((1 - ncc_epsilon) * lower +
+                                                 ncc_epsilon / len(self.feature_count[count_string]))
+                    upper = (self.feature_count[count_string][f_val_index]) / (num_items + ncc_s_param)
+                    u_numerator = u_numerator * ((1 - ncc_epsilon) * upper +
+                                                 ncc_epsilon / len(self.feature_count[count_string]))
+                resulting_score[j, 1] = u_denom / (u_denom + u_numerator)
+                resulting_score[j, 0] = l_denom / (l_denom + l_numerator)
+            result = Scores(resulting_score)
             answers.append(result)
-        
+
         return answers
-
