+++ examples/multilabel/mlchaintest.py
@@ -6,11 +6,11 @@ Created on 20 may 2019
 About how to use the Imprecise multilabel chaining
 '''
 from classifip.dataset import arff
-from classifip.models.mlcncc import MLCNCCOuterApprox, MLCNCCExact
+from classifip.models.mlc.chainncc import MLChaining
 import timeit
 
 dataArff = arff.ArffFile()
-dataArff.load("labels5.arff")
+dataArff.load("emotions.arff")
 dataArff.discretize(discmet='eqfreq', numint=5)
 nb_labels = 5
 
@@ -18,7 +18,7 @@ nb_labels = 5
 new_instances = [row[0:len(row) - nb_labels] for row in dataArff.data[10:11]]
 
 # We start by creating a model
-model = MLCNCCOuterApprox()
+model = MLChaining()
 model.learn(dataArff, nb_labels)  # , seed_random_label=134)
 
 probabilities, chain = model.evaluate(new_instances, ncc_epsilon=0.001, ncc_s_param=1)
@@ -28,18 +28,4 @@ print(probabilities[0])
 
 print("Label true", dataArff.data[0:1][0][-nb_labels:], '\n')
 print("Label predicts: ", chain, '\n')
-
-# Inference exact with complexity minimal
-model = MLCNCCExact()
-model.learn(dataArff, nb_labels=nb_labels)
-
-start = timeit.default_timer()
-solution_exact_1 = model.evaluate(new_instances, ncc_epsilon=0.001, ncc_s_param=1)
-print('Solution with minimum complexity:', solution_exact_1, timeit.default_timer() - start)
-start = timeit.default_timer()
-solution_exact_2 = model.evaluate_exact(new_instances, ncc_epsilon=0.001, ncc_s_param=1)
-print('Solution with maximum complexity:', solution_exact_2, timeit.default_timer() - start)
-
-# exec('def foo(a, b):  return a + b')
-# print(foo(1, 2))
